+++ frontend/src/routes/problems/$problemId/-components/code-panel/code-panel.tsx
@@ -2,7 +2,7 @@ import type { ComponentPropsWithoutRef, FC } from "react"
 
 import { cn } from "@/lib/utils"
 
-import { useProblem } from "../../-contexts/problem-context/problem-context"
+import { useProblem } from "../../-contexts/problem-context"
 import { Editor } from "./editor"
 import { LanguageSelector } from "./language-selector"
 import { SubmitButton } from "./submit-button"
