+++ frontend/src/routes/problems/$problemId/-contexts/problem-context/index.ts
@@ -1,0 +1,8 @@
+export {
+  type Language,
+  type Problem,
+  type ProblemContext,
+  ProblemProvider,
+  type ProblemProviderProps,
+  useProblem,
+} from "./problem-context"
